move methods with signatures not supported by prettier into their own files
