+++ CopilotKit/packages/backend/src/lib/copilotkit-backend-implementation.ts
@@ -1,0 +1,150 @@
+import http from "http";
+import {
+  Action,
+  ToolDefinition,
+  EXCLUDE_FROM_FORWARD_PROPS_KEYS,
+  actionToChatCompletionFunction,
+  Parameter,
+} from "@copilotkit/shared";
+import { copilotkitStreamInterceptor, remoteChainToAction } from "../utils";
+import { RemoteChain, CopilotKitServiceAdapter } from "../types";
+
+interface CopilotBackendConstructorParams {
+  actions?: Action<any>[];
+  langserve?: RemoteChain[];
+  debug?: boolean;
+}
+
+export class CopilotBackendImplementation {
+  private actions: Action<any>[] = [];
+  private langserve: Promise<Action<any>>[] = [];
+  private debug: boolean = false;
+
+  constructor(params?: CopilotBackendConstructorParams) {
+    for (const action of params?.actions || []) {
+      this.actions.push(action);
+    }
+    for (const chain of params?.langserve || []) {
+      this.langserve.push(remoteChainToAction(chain));
+    }
+    this.debug = params?.debug || false;
+  }
+
+  addAction(action: Action<any>): void {
+    this.removeAction(action.name);
+    this.actions.push(action);
+  }
+
+  removeAction(actionName: string): void {
+    this.actions = this.actions.filter((f) => f.name !== actionName);
+  }
+
+  removeBackendOnlyProps(forwardedProps: any): void {
+    this.removeBackendOnlyProps(forwardedProps);
+    // Get keys backendOnlyPropsKeys in order to remove them from the forwardedProps
+    const backendOnlyPropsKeys = forwardedProps[EXCLUDE_FROM_FORWARD_PROPS_KEYS];
+    if (Array.isArray(backendOnlyPropsKeys)) {
+      backendOnlyPropsKeys.forEach((key) => {
+        const success = Reflect.deleteProperty(forwardedProps, key);
+        if (!success) {
+          console.error(`Failed to delete property ${key}`);
+        }
+      });
+      // After deleting individual backend-only properties, delete the EXCLUDE_FROM_FORWARD_PROPS_KEYS property itself from forwardedProps
+      const success = Reflect.deleteProperty(forwardedProps, EXCLUDE_FROM_FORWARD_PROPS_KEYS);
+      if (!success) {
+        console.error(`Failed to delete EXCLUDE_FROM_FORWARD_PROPS_KEYS`);
+      }
+    } else if (backendOnlyPropsKeys) {
+      console.error("backendOnlyPropsKeys is not an array");
+    }
+  }
+  async stream(
+    forwardedProps: any,
+    serviceAdapter: CopilotKitServiceAdapter,
+  ): Promise<ReadableStream> {
+    const langserveFunctions: Action<any>[] = [];
+
+    for (const chainPromise of this.langserve) {
+      try {
+        const chain = await chainPromise;
+        langserveFunctions.push(chain);
+      } catch (error) {
+        console.error("Error loading langserve chain:", error);
+      }
+    }
+
+    // merge server side functions with langserve functions
+    let mergedTools = mergeServerSideTools(
+      this.actions.map(actionToChatCompletionFunction),
+      langserveFunctions.map(actionToChatCompletionFunction),
+    );
+
+    // merge with client side functions
+    mergedTools = mergeServerSideTools(mergedTools, forwardedProps.tools);
+
+    const openaiCompatibleStream = await serviceAdapter.stream({
+      ...forwardedProps,
+      tools: mergedTools,
+    });
+    return copilotkitStreamInterceptor(
+      openaiCompatibleStream,
+      [...this.actions, ...langserveFunctions],
+      this.debug,
+    );
+  }
+
+  async response(req: Request, serviceAdapter: CopilotKitServiceAdapter): Promise<Response> {
+    try {
+      return new Response(await this.stream(await req.json(), serviceAdapter));
+    } catch (error: any) {
+      return new Response("", { status: 500, statusText: error.message });
+    }
+  }
+
+  async streamHttpServerResponse(
+    req: http.IncomingMessage,
+    res: http.ServerResponse,
+    serviceAdapter: CopilotKitServiceAdapter,
+  ) {
+    const bodyParser = new Promise<any>((resolve, reject) => {
+      let body = "";
+      req.on("data", (chunk) => (body += chunk.toString()));
+      req.on("end", () => {
+        try {
+          resolve(JSON.parse(body));
+        } catch (error) {
+          reject(error);
+        }
+      });
+    });
+    const forwardedProps = await bodyParser;
+    const stream = await this.stream(forwardedProps, serviceAdapter);
+    const reader = stream.getReader();
+
+    while (true) {
+      const { done, value } = await reader.read();
+      if (done) {
+        res.end();
+        break;
+      } else {
+        res.write(new TextDecoder().decode(value));
+      }
+    }
+  }
+}
+
+export function mergeServerSideTools(
+  serverTools: ToolDefinition[],
+  clientTools?: ToolDefinition[],
+) {
+  let allTools: ToolDefinition[] = serverTools.slice();
+  const serverToolsNames = serverTools.map((tool) => tool.function.name);
+  if (clientTools) {
+    allTools = allTools.concat(
+      // filter out any client functions that are already defined on the server
+      clientTools.filter((tool: ToolDefinition) => !serverToolsNames.includes(tool.function.name)),
+    );
+  }
+  return allTools;
+}
